tried to meet the functional request in this homework in O(n^2) without testing

+++ src/Core/ray.hpp
@@ -1,15 +1,15 @@
 #pragma once
 
 struct Ray {
-    long long x1, y1, x2, y2;
-    long long dx, dy, x2y1, x1y2, x2y1_x1y2;
+    int x1, y1, x2, y2;
+    int dx, dy, x2y1, x1y2, x2y1_x1y2;
     Ray(const int x1, const int y1, const int x2, const int y2) noexcept
         : x1(x1), y1(y1), x2(x2), y2(y2)
     {
-        dx = (long long)x1 - (long long)x2;
-        dy = (long long)y1 - (long long)y2;
-        x2y1 = (long long)x2 * (long long)y1;
-        x1y2 = (long long)x1 * (long long)y2;
+        dx = (int)x1 - (int)x2;
+        dy = (int)y1 - (int)y2;
+        x2y1 = (int)x2 * (int)y1;
+        x1y2 = (int)x1 * (int)y2;
         x2y1_x1y2 = x2y1 - x1y2;
     }
 };
